pyproject.toml
[tool.poetry]
name = "schemarrow"
version = "0.1.0"
description = "A library for switching pandas backend to pyarrow"
authors = ["DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/DanielAvdar/SchemArrow"
repository = "https://github.com/DanielAvdar/SchemArrow"
documentation = "https://github.com/DanielAvdar/SchemArrow"
keywords = [
  "pandas",
  "pyarrow",
  "arrow",
  "schemarrow",
  "pandas-pyarrow",
  "pandas-arrow",
  "pandas-backend",
  "data-processing",
  "data-conversion",
  "backend-switch",
  "high-performance",
  "datasets",
  "big-data",
  "data-serialization",
  "data-analysis",
  "data-science",
  "machine-learning",
  "AI",
  "python-library",
  "data-backend",
  "performance-optimization",
  "memory-efficient",
  "python"
]
packages = [ { include = "schemarrow", from = "." } ]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pandas = ">=2"
pyarrow = ">=7.0.0, <=15.0.0"
db-dtypes = { extras = ["db-dtypes"], version = ">=1", optional = true }

[tool.poetry.dev-dependencies]
pytest = ">= 7.1.2"
hypothesis = ">= 6.23.3"
pre-commit = ">= 2.20.0"
pytest-cov = ">= 4.0.0"
mypy = ">= 0.991"
pytest-parametrization = "^2022"
pandas-gbq = "*"
types-pytz = "*"
pandas-stubs = "*"
ruff = "^0.3.1"

[tool.poetry.extras]
db-dtypes = ["db-dtypes"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
include = ["pyproject.toml", "tests/**",  "schemarrow/**",]# ".github/workflows/**", ]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "I"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
combine-as-imports = true
section-order = ["future", "standard-library", "first-party", "local-folder", "third-party"]
[tool.mypy]
python_version = 3.11
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = false
warn_no_return = false
warn_unreachable = false
exclude = ["^test/"]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true
